[Refactor] Update type casting methods to use TypeManager for improved consistency and maintainability
