+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/expression/DateLiteral.java
@@ -41,6 +41,7 @@ import com.starrocks.sql.ast.AstVisitor;
 import com.starrocks.sql.ast.AstVisitorExtendInterface;
 import com.starrocks.sql.common.ErrorType;
 import com.starrocks.sql.common.StarRocksPlannerException;
+import com.starrocks.sql.common.TypeManager;
 import com.starrocks.type.PrimitiveType;
 import com.starrocks.type.Type;
 import com.starrocks.type.TypeFactory;
@@ -303,7 +304,7 @@ public class DateLiteral extends LiteralExpr {
             }
         } else if (targetType.isStringType()) {
             return new StringLiteral(getStringValue());
-        } else if (Type.isImplicitlyCastable(this.type, targetType, true)) {
+        } else if (TypeManager.isImplicitlyCastable(this.type, targetType, true)) {
             return new CastExpr(targetType, this);
         }
         Preconditions.checkState(false);
