+++ fe/fe-core/src/main/java/com/starrocks/sql/common/TypeManager.java
@@ -29,6 +29,7 @@ import com.starrocks.type.ScalarType;
 import com.starrocks.type.StructField;
 import com.starrocks.type.StructType;
 import com.starrocks.type.Type;
+import com.starrocks.type.TypeCompatibilityMatrix;
 import com.starrocks.type.TypeFactory;
 
 import java.util.ArrayList;
@@ -83,7 +84,7 @@ public class TypeManager {
     }
 
     private static Type getCommonScalarType(ScalarType t1, ScalarType t2) {
-        return ScalarType.getAssignmentCompatibleType(t1, t2, false);
+        return getAssignmentCompatibleType(t1, t2, false);
     }
 
     private static Type getCommonArrayType(ArrayType t1, ArrayType t2) {
@@ -128,7 +129,7 @@ public class TypeManager {
                 return expr;
             }
 
-            if (!Type.canCastTo(expr.getType(), targetType)) {
+            if (!canCastTo(expr.getType(), targetType)) {
                 throw new SemanticException(
                         "Cannot cast '" + ExprToSql.toSql(expr) + "' from " + expr.getType() + " to " + targetType);
             }
@@ -178,7 +179,7 @@ public class TypeManager {
             Type decimalType = t1.isDecimalV3() ? t1 : t2;
             Type otherType = t1.isDecimalV3() ? t2 : t1;
             Preconditions.checkState(otherType.isScalarType());
-            return ScalarType.getAssigmentCompatibleTypeOfDecimalV3((ScalarType) decimalType, (ScalarType) otherType);
+            return getAssigmentCompatibleTypeOfDecimalV3((ScalarType) decimalType, (ScalarType) otherType);
         }
 
         if (t1.getPrimitiveType() != PrimitiveType.INVALID_TYPE &&
@@ -198,7 +199,7 @@ public class TypeManager {
             return Type.VARCHAR;
         }
         if (t1ResultType == PrimitiveType.BIGINT && t2ResultType == PrimitiveType.BIGINT) {
-            return Type.getAssignmentCompatibleType(t1, t2, false);
+            return getAssignmentCompatibleType(t1, t2, false);
         }
 
         if ((t1ResultType == PrimitiveType.BIGINT
@@ -257,7 +258,7 @@ public class TypeManager {
             Type decimalType = type1.isDecimalV3() ? type1 : type2;
             Type otherType = type1.isDecimalV3() ? type2 : type1;
             Preconditions.checkState(otherType.isScalarType());
-            return ScalarType.getAssigmentCompatibleTypeOfDecimalV3((ScalarType) decimalType, (ScalarType) otherType);
+            return getAssigmentCompatibleTypeOfDecimalV3((ScalarType) decimalType, (ScalarType) otherType);
         }
 
         BiFunction<Type, Type, Boolean> isDataAndString =
@@ -331,5 +332,323 @@ public class TypeManager {
         } else {
             return compatibleType;
         }
+    }
+
+    public static boolean canCastTo(Type from, Type to) {
+        if (from.isNull()) {
+            return true;
+        } else if (from.isStringType() && to.isBitmapType()) {
+            return true;
+        } else if (from.isScalarType() && to.isScalarType()) {
+            return canCastTo((ScalarType) from, (ScalarType) to);
+        } else if (from.isArrayType() && to.isArrayType()) {
+            return canCastTo(((ArrayType) from).getItemType(), ((ArrayType) to).getItemType());
+        } else if (from.isMapType() && to.isMapType()) {
+            MapType fromMap = (MapType) from;
+            MapType toMap = (MapType) to;
+            return canCastTo(fromMap.getKeyType(), toMap.getKeyType()) &&
+                    canCastTo(fromMap.getValueType(), toMap.getValueType());
+        } else if (from.isStructType() && to.isStructType()) {
+            StructType fromStruct = (StructType) from;
+            StructType toStruct = (StructType) to;
+            if (fromStruct.getFields().size() != toStruct.getFields().size()) {
+                return false;
+            }
+            for (int i = 0; i < fromStruct.getFields().size(); ++i) {
+                if (!canCastTo(fromStruct.getField(i).getType(), toStruct.getField(i).getType())) {
+                    return false;
+                }
+            }
+            return true;
+        } else if (from.isStringType() && to.isArrayType()) {
+            return true;
+        } else if (from.isJsonType() && to.isArrayType()) {
+            ArrayType array = (ArrayType) to;
+            if (array.getItemType().isScalarType() || array.getItemType().isStructType()) {
+                return true;
+            }
+            return false;
+        } else if (from.isJsonType() && to.isStructType()) {
+            return true;
+        } else if (from.isJsonType() && to.isMapType()) {
+            MapType map = (MapType) to;
+            return canCastTo(Type.VARCHAR, map.getKeyType()) && canCastTo(Type.JSON, map.getValueType());
+        } else if (from.isBoolean() && to.isComplexType()) {
+            // for mock nest type with NULL value, the cast must return NULL
+            // like cast(map{1: NULL} as MAP<int, int>)
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * Return type t such that values from both t1 and t2 can be assigned to t without an
+     * explicit cast. If strict, does not consider conversions that would result in loss
+     * of precision (e.g. converting decimal to float). Returns INVALID_TYPE if there is
+     * no such type or if any of t1 and t2 is INVALID_TYPE.
+     * TODO: Support non-scalar types.
+     */
+    public static Type getAssignmentCompatibleType(Type t1, Type t2, boolean strict) {
+        if (t1.isScalarType() && t2.isScalarType()) {
+            return getAssignmentCompatibleType((ScalarType) t1, (ScalarType) t2, strict);
+        }
+        return ScalarType.INVALID;
+    }
+
+    public static Type getCommonType(Type[] argTypes, int fromIndex, int toIndex) {
+        Preconditions.checkState(argTypes != null);
+        Preconditions.checkState(0 <= fromIndex && fromIndex < toIndex && toIndex <= argTypes.length);
+        Type commonType = argTypes[fromIndex];
+        for (int i = fromIndex + 1; i < toIndex; ++i) {
+            commonType = getCommonType(commonType, argTypes[i]);
+        }
+        return commonType;
+    }
+
+    /**
+     * Given two types, return the common supertype of them.
+     *
+     * @return the common type, INVALID if no common type exists.
+     */
+    public static Type getCommonType(Type t1, Type t2) {
+        if (t1.isScalarType() && t2.isScalarType()) {
+            return getAssignmentCompatibleType((ScalarType) t1, (ScalarType) t2, true);
+        }
+        if (t1.isArrayType() && t2.isArrayType()) {
+            ArrayType arrayType1 = (ArrayType) t1;
+            ArrayType arrayType2 = (ArrayType) t2;
+
+            Type item1 = arrayType1.getItemType();
+            Type item2 = arrayType2.getItemType();
+            Type common = getCommonType(item1, item2);
+            return common.isValid() ? new ArrayType(common) : common;
+        }
+        if (t1.isNull() || t2.isNull()) {
+            return t1.isNull() ? t2 : t1;
+        }
+        return Type.INVALID;
+    }
+
+    // getAssigmentCompatibleTypeOfDecimalV3 is used by FunctionCallExpr for finding the common Type of argument types.
+    // i.e.
+    // least(decimal64(15,2), decimal32(5,4)) match least instance: least(decimal64(18,4), decimal64(18,4));
+    // least(double, decimal128(38,7)) match least instance: least(double, double).
+    public static ScalarType getAssigmentCompatibleTypeOfDecimalV3(
+            ScalarType decimalType, ScalarType otherType) {
+        Preconditions.checkState(decimalType.isDecimalV3());
+        // same triple (type, precision, scale)
+        if (decimalType.equals(otherType)) {
+            PrimitiveType type = PrimitiveType.getWiderDecimalV3Type(PrimitiveType.DECIMAL64, decimalType.getType());
+            return TypeFactory.createDecimalV3Type(type, decimalType.getScalarPrecision(), decimalType.getScalarScale());
+        }
+
+        switch (otherType.getPrimitiveType()) {
+            case DECIMALV2:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 27, 9));
+            case DECIMAL32:
+            case DECIMAL64:
+            case DECIMAL128:
+            case DECIMAL256:
+                return getCommonTypeForDecimalV3(decimalType, otherType);
+
+            case BOOLEAN:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 1, 0));
+            case TINYINT:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 3, 0));
+            case SMALLINT:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 5, 0));
+            case INT:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL64, 10, 0));
+            case BIGINT:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 19, 0));
+            case LARGEINT:
+                return getCommonTypeForDecimalV3(decimalType,
+                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 38, 0));
+            case FLOAT:
+            case DOUBLE:
+            case CHAR:
+            case VARCHAR:
+            case DATE:
+            case DATETIME:
+            case TIME:
+            case JSON:
+                return Type.DOUBLE;
+            default:
+                return Type.INVALID;
+        }
+    }
+
+    // isAssignable means that assigning or casting rhs to lhs never overflows.
+    // only both integer part width and fraction part width of lhs is not narrower than counterparts
+    // of rhs, then rhs can be assigned to lhs. for integer types, integer part width is computed by
+    // calling Type::getPrecision and its scale is 0.
+    public static boolean isAssignable2Decimal(ScalarType lhs, ScalarType rhs) {
+        int lhsIntPartWidth;
+        int lhsScale;
+        int rhsIntPartWidth;
+        int rhsScale;
+        if (lhs.isFixedPointType()) {
+            lhsIntPartWidth = lhs.getPrecision();
+            lhsScale = 0;
+        } else {
+            lhsIntPartWidth = lhs.getScalarPrecision() - lhs.getScalarScale();
+            lhsScale = lhs.getScalarScale();
+        }
+
+        if (rhs.isFixedPointType()) {
+            rhsIntPartWidth = rhs.getPrecision();
+            rhsScale = 0;
+        } else {
+            rhsIntPartWidth = rhs.getScalarPrecision() - rhs.getScalarScale();
+            rhsScale = rhs.getScalarScale();
+        }
+
+        // when lhs is integer, for instance, tinyint, lhsIntPartWidth is 3, it cannot holds
+        // a DECIMAL(3, 0).
+        if (lhs.isFixedPointType() && rhs.isDecimalOfAnyVersion()) {
+            return lhsIntPartWidth > rhsIntPartWidth && lhsScale >= rhsScale;
+        } else {
+            return lhsIntPartWidth >= rhsIntPartWidth && lhsScale >= rhsScale;
+        }
+    }
+
+    /**
+     * Return type t such that values from both t1 and t2 can be assigned to t.
+     * If strict, only return types when there will be no loss of precision.
+     * Returns INVALID_TYPE if there is no such type or if any of t1 and t2
+     * is INVALID_TYPE.
+     */
+    public static ScalarType getAssignmentCompatibleType(ScalarType t1, ScalarType t2, boolean strict) {
+        if (!t1.isValid() || !t2.isValid()) {
+            return Type.INVALID;
+        }
+        if (t1.isUnknown() || t2.isUnknown()) {
+            return Type.UNKNOWN_TYPE;
+        }
+        if (t1.equals(t2)) {
+            return t1;
+        }
+        if (t1.isNull()) {
+            return t2;
+        }
+        if (t2.isNull()) {
+            return t1;
+        }
+
+        boolean t1IsHLL = t1.getType() == PrimitiveType.HLL;
+        boolean t2IsHLL = t2.getType() == PrimitiveType.HLL;
+        if (t1IsHLL || t2IsHLL) {
+            if (t1IsHLL && t2IsHLL) {
+                return TypeFactory.createHllType();
+            }
+            return Type.INVALID;
+        }
+
+        if (t1.isStringType() || t2.isStringType()) {
+            return TypeFactory.createVarcharType(Math.max(t1.getLength(), t2.getLength()));
+        }
+
+        if (t1.isDecimalV3()) {
+            return getAssigmentCompatibleTypeOfDecimalV3(t1, t2);
+        }
+
+        if (t2.isDecimalV3()) {
+            return getAssigmentCompatibleTypeOfDecimalV3(t2, t1);
+        }
+
+        if (t1.isDecimalOfAnyVersion() && t2.isDate() || t1.isDate() && t2.isDecimalOfAnyVersion()) {
+            return Type.INVALID;
+        }
+
+        if (t1.isDecimalV2() || t2.isDecimalV2()) {
+            return Type.DECIMALV2;
+        }
+
+        if (t1.isFunctionType() || t2.isFunctionType()) {
+            return Type.INVALID;
+        }
+
+        PrimitiveType smallerType =
+                (t1.getType().ordinal() < t2.getType().ordinal() ? t1.getType() : t2.getType());
+        PrimitiveType largerType =
+                (t1.getType().ordinal() > t2.getType().ordinal() ? t1.getType() : t2.getType());
+        PrimitiveType result = TypeCompatibilityMatrix.getCompatibleType(smallerType, largerType);
+        Preconditions.checkNotNull(result, String.format("No assignment from %s to %s", t1, t2));
+        return TypeFactory.createType(result);
+    }
+
+    /**
+     * Returns true if t1 can be implicitly cast to t2 according to Impala's casting rules.
+     * Implicit casts are always allowed when no loss of precision would result (i.e. every
+     * value of t1 can be represented exactly by a value of t2). Implicit casts are allowed
+     * in certain other cases such as casting numeric types to floating point types and
+     * converting strings to timestamps.
+     * If strict is true, only consider casts that result in no loss of precision.
+     * TODO: Support casting of non-scalar types.
+     */
+    public static boolean isImplicitlyCastable(Type t1, Type t2, boolean strict) {
+        if (t1.isScalarType() && t2.isScalarType()) {
+            return isImplicitlyCastable((ScalarType) t1, (ScalarType) t2, strict);
+        }
+        if (t1.isArrayType() && t2.isArrayType()) {
+            return isImplicitlyCastable(((ArrayType) t1).getItemType(), ((ArrayType) t2).getItemType(), strict);
+        }
+        return false;
+    }
+
+    // A common type for two decimal v3 types means that if t2 = getCommonTypeForDecimalV3(t0, t1),
+    // two invariants following is always holds:
+    // 1. t2's integer part is sufficient to hold both t0 and t1's counterparts: i.e.
+    //    t2.precision - t2.scale = max((t0.precision - t0.scale), (t1.precision - t1.scale))
+    // 2. t2's fraction part is sufficient to hold both t0 and t1's counterparts: i.e.
+    //    t2.scale = max(t0.scale, t1.scale)
+    public static ScalarType getCommonTypeForDecimalV3(ScalarType lhs, ScalarType rhs) {
+        Preconditions.checkState(lhs.isDecimalV3() && rhs.isDecimalV3());
+        int lhsPrecision = lhs.getScalarPrecision();
+        int lhsScale = lhs.getScalarScale();
+        int rhsPrecision = rhs.getScalarPrecision();
+        int rhsScale = rhs.getScalarScale();
+        int lhsIntegerPartWidth = lhsPrecision - lhsScale;
+        int rhsIntegerPartWidth = rhsPrecision - rhsScale;
+        int integerPartWidth = Math.max(lhsIntegerPartWidth, rhsIntegerPartWidth);
+        int scale = Math.max(lhsScale, rhsScale);
+        int precision = integerPartWidth + scale;
+        boolean hasDecimal256 = lhs.isDecimal256() || rhs.isDecimal256();
+        // TODO(stephen): support auto scale up decimal precision
+        if ((precision > 38 && !hasDecimal256) || (precision > 76)) {
+            return ScalarType.DOUBLE;
+        } else {
+            // the common type's PrimitiveType of two decimal types should wide enough, i.e
+            // the common type of (DECIMAL32, DECIMAL64) should be DECIMAL64
+            PrimitiveType primitiveType =
+                    PrimitiveType.getWiderDecimalV3Type(lhs.getPrimitiveType(), rhs.getPrimitiveType());
+            // the narrowestType for specified precision and scale is just wide properly to hold a decimal value, i.e
+            // DECIMAL128(7,4), DECIMAL64(7,4) and DECIMAL32(7,4) can all be held in a DECIMAL32(7,4) type without
+            // precision loss.
+            Type narrowestType = TypeFactory.createDecimalV3NarrowestType(precision, scale);
+            primitiveType = PrimitiveType.getWiderDecimalV3Type(primitiveType, narrowestType.getPrimitiveType());
+            // create a commonType with wider primitive type.
+            return TypeFactory.createDecimalV3Type(primitiveType, precision, scale);
+        }
+    }
+
+    /**
+     * Returns true t1 can be implicitly cast to t2, false otherwise.
+     * If strict is true, only consider casts that result in no loss of precision.
+     */
+    public static boolean isImplicitlyCastable(ScalarType t1, ScalarType t2, boolean strict) {
+        return getAssignmentCompatibleType(t1, t2, strict).matchesType(t2);
+    }
+
+    // TODO(mofei) Why call implicit cast in the explicit cast context
+    public static boolean canCastTo(ScalarType type, ScalarType targetType) {
+        return PrimitiveType.isImplicitCast(type.getPrimitiveType(), targetType.getPrimitiveType());
     }
 }
