+++ fe/fe-core/src/main/java/com/starrocks/type/ScalarType.java
@@ -88,173 +88,8 @@ public class ScalarType extends Type implements Cloneable {
         this.type = PrimitiveType.INVALID_TYPE;
     }
 
-    // A common type for two decimal v3 types means that if t2 = getCommonTypeForDecimalV3(t0, t1),
-    // two invariants following is always holds:
-    // 1. t2's integer part is sufficient to hold both t0 and t1's counterparts: i.e.
-    //    t2.precision - t2.scale = max((t0.precision - t0.scale), (t1.precision - t1.scale))
-    // 2. t2's fraction part is sufficient to hold both t0 and t1's counterparts: i.e.
-    //    t2.scale = max(t0.scale, t1.scale)
-    public static ScalarType getCommonTypeForDecimalV3(ScalarType lhs, ScalarType rhs) {
-        Preconditions.checkState(lhs.isDecimalV3() && rhs.isDecimalV3());
-        int lhsPrecision = lhs.getScalarPrecision();
-        int lhsScale = lhs.getScalarScale();
-        int rhsPrecision = rhs.getScalarPrecision();
-        int rhsScale = rhs.getScalarScale();
-        int lhsIntegerPartWidth = lhsPrecision - lhsScale;
-        int rhsIntegerPartWidth = rhsPrecision - rhsScale;
-        int integerPartWidth = Math.max(lhsIntegerPartWidth, rhsIntegerPartWidth);
-        int scale = Math.max(lhsScale, rhsScale);
-        int precision = integerPartWidth + scale;
-        boolean hasDecimal256 = lhs.isDecimal256() || rhs.isDecimal256();
-        // TODO(stephen): support auto scale up decimal precision
-        if ((precision > 38 && !hasDecimal256) || (precision > 76)) {
-            return ScalarType.DOUBLE;
-        } else {
-            // the common type's PrimitiveType of two decimal types should wide enough, i.e
-            // the common type of (DECIMAL32, DECIMAL64) should be DECIMAL64
-            PrimitiveType primitiveType =
-                    PrimitiveType.getWiderDecimalV3Type(lhs.getPrimitiveType(), rhs.getPrimitiveType());
-            // the narrowestType for specified precision and scale is just wide properly to hold a decimal value, i.e
-            // DECIMAL128(7,4), DECIMAL64(7,4) and DECIMAL32(7,4) can all be held in a DECIMAL32(7,4) type without
-            // precision loss.
-            Type narrowestType = TypeFactory.createDecimalV3NarrowestType(precision, scale);
-            primitiveType = PrimitiveType.getWiderDecimalV3Type(primitiveType, narrowestType.getPrimitiveType());
-            // create a commonType with wider primitive type.
-            return TypeFactory.createDecimalV3Type(primitiveType, precision, scale);
-        }
-    }
-
-    // getAssigmentCompatibleTypeOfDecimalV3 is used by FunctionCallExpr for finding the common Type of argument types.
-    // i.e.
-    // least(decimal64(15,2), decimal32(5,4)) match least instance: least(decimal64(18,4), decimal64(18,4));
-    // least(double, decimal128(38,7)) match least instance: least(double, double).
-    public static ScalarType getAssigmentCompatibleTypeOfDecimalV3(
-            ScalarType decimalType, ScalarType otherType) {
-        Preconditions.checkState(decimalType.isDecimalV3());
-        // same triple (type, precision, scale)
-        if (decimalType.equals(otherType)) {
-            PrimitiveType type = PrimitiveType.getWiderDecimalV3Type(PrimitiveType.DECIMAL64, decimalType.type);
-            return TypeFactory.createDecimalV3Type(type, decimalType.precision, decimalType.scale);
-        }
-
-        switch (otherType.getPrimitiveType()) {
-            case DECIMALV2:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 27, 9));
-            case DECIMAL32:
-            case DECIMAL64:
-            case DECIMAL128:
-            case DECIMAL256:
-                return getCommonTypeForDecimalV3(decimalType, otherType);
-
-            case BOOLEAN:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 1, 0));
-            case TINYINT:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 3, 0));
-            case SMALLINT:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL32, 5, 0));
-            case INT:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL64, 10, 0));
-            case BIGINT:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 19, 0));
-            case LARGEINT:
-                return getCommonTypeForDecimalV3(decimalType,
-                        TypeFactory.createDecimalV3Type(PrimitiveType.DECIMAL128, 38, 0));
-            case FLOAT:
-            case DOUBLE:
-            case CHAR:
-            case VARCHAR:
-            case DATE:
-            case DATETIME:
-            case TIME:
-            case JSON:
-                return DOUBLE;
-            default:
-                return INVALID;
-        }
-    }
-
-    /**
-     * Return type t such that values from both t1 and t2 can be assigned to t.
-     * If strict, only return types when there will be no loss of precision.
-     * Returns INVALID_TYPE if there is no such type or if any of t1 and t2
-     * is INVALID_TYPE.
-     */
-    public static ScalarType getAssignmentCompatibleType(ScalarType t1, ScalarType t2, boolean strict) {
-        if (!t1.isValid() || !t2.isValid()) {
-            return INVALID;
-        }
-        if (t1.isUnknown() || t2.isUnknown()) {
-            return UNKNOWN_TYPE;
-        }
-        if (t1.equals(t2)) {
-            return t1;
-        }
-        if (t1.isNull()) {
-            return t2;
-        }
-        if (t2.isNull()) {
-            return t1;
-        }
-
-        boolean t1IsHLL = t1.type == PrimitiveType.HLL;
-        boolean t2IsHLL = t2.type == PrimitiveType.HLL;
-        if (t1IsHLL || t2IsHLL) {
-            if (t1IsHLL && t2IsHLL) {
-                return TypeFactory.createHllType();
-            }
-            return INVALID;
-        }
-
-        if (t1.isStringType() || t2.isStringType()) {
-            return TypeFactory.createVarcharType(Math.max(t1.len, t2.len));
-        }
-
-        if (t1.isDecimalV3()) {
-            return getAssigmentCompatibleTypeOfDecimalV3(t1, t2);
-        }
-
-        if (t2.isDecimalV3()) {
-            return getAssigmentCompatibleTypeOfDecimalV3(t2, t1);
-        }
-
-        if (t1.isDecimalOfAnyVersion() && t2.isDate() || t1.isDate() && t2.isDecimalOfAnyVersion()) {
-            return INVALID;
-        }
-
-        if (t1.isDecimalV2() || t2.isDecimalV2()) {
-            return DECIMALV2;
-        }
-
-        if (t1.isFunctionType() || t2.isFunctionType()) {
-            return INVALID;
-        }
-
-        PrimitiveType smallerType =
-                (t1.type.ordinal() < t2.type.ordinal() ? t1.type : t2.type);
-        PrimitiveType largerType =
-                (t1.type.ordinal() > t2.type.ordinal() ? t1.type : t2.type);
-        PrimitiveType result = TypeCompatibilityMatrix.getCompatibleType(smallerType, largerType);
-        Preconditions.checkNotNull(result, String.format("No assignment from %s to %s", t1, t2));
-        return TypeFactory.createType(result);
-    }
-
-    /**
-     * Returns true t1 can be implicitly cast to t2, false otherwise.
-     * If strict is true, only consider casts that result in no loss of precision.
-     */
-    public static boolean isImplicitlyCastable(ScalarType t1, ScalarType t2, boolean strict) {
-        return getAssignmentCompatibleType(t1, t2, strict).matchesType(t2);
-    }
-
-    // TODO(mofei) Why call implicit cast in the explicit cast context
-    public static boolean canCastTo(ScalarType type, ScalarType targetType) {
-        return PrimitiveType.isImplicitCast(type.getPrimitiveType(), targetType.getPrimitiveType());
+    public PrimitiveType getType() {
+        return type;
     }
 
     @Override
